backend-moderation-platform/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
apscheduler
python-dotenv
PyJWT